refactor: introduce option for node filtering
